requirements.txt
numpy==1.24.2
opencv-python==4.7.0.68
pyserial==3.5
requests==2.28.2
pygame==2.2.0
pytest==7.2.2
onnxruntime==1.14.1
pylint
ultralytics==8.0.53
keyboard==0.13.5
ipykernel
face-recognition
pylint
